refactor: #31 マイグレーション統合とUser.skills削除
- User.skillsカラム削除（SkillTreeテーブルに移行）
- マイグレーションを7本から1本に統合（issue単位一括）
- ADR 008更新: MVP優先でissue単位一括マイグレーションに変更

理由：
- SkillTreeテーブル実装によりskillsカラムは冗長
- 部分的ロールバックの必要性が薄く、all-or-nothingが適切
- マイグレーションの本質（環境同期）に集中
