+++ backend/alembic/versions/0001_create_all_tables.py
@@ -1,0 +1,185 @@
+"""create all tables (issue #27 + #31)
+
+Revision ID: 0001_all_tables
+Revises:
+Create Date: 2026-02-19
+
+"""
+
+from typing import Sequence, Union
+
+import sqlalchemy as sa
+from alembic import op
+
+revision: str = "0001_all_tables"
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """全テーブルを一括作成（issue #27 + #31）"""
+
+    # 1. usersテーブル（issue #27、skillsカラムなし）
+    op.create_table(
+        "users",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("username", sa.String(), nullable=False),
+        sa.Column("level", sa.Integer(), nullable=True),
+        sa.Column("exp", sa.Integer(), nullable=True),
+        sa.Column("rank", sa.Integer(), nullable=True),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=True,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=True,
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_users_id"), "users", ["id"], unique=False)
+    op.create_index(op.f("ix_users_username"), "users", ["username"], unique=True)
+
+    # 2. profilesテーブル（issue #31）
+    op.create_table(
+        "profiles",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("user_id", sa.Integer(), nullable=False),
+        sa.Column("github_username", sa.String(), nullable=True),
+        sa.Column("qiita_id", sa.String(), nullable=True),
+        sa.Column("connpass_id", sa.String(), nullable=True),
+        sa.Column("portfolio_url", sa.String(), nullable=True),
+        sa.Column("portfolio_text", sa.Text(), nullable=True),
+        sa.Column("last_analyzed_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"]),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("user_id"),
+    )
+    op.create_index(op.f("ix_profiles_user_id"), "profiles", ["user_id"], unique=True)
+
+    # 3. oauth_accountsテーブル（issue #31）
+    op.create_table(
+        "oauth_accounts",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("user_id", sa.Integer(), nullable=False),
+        sa.Column("provider", sa.String(), nullable=False),
+        sa.Column("provider_user_id", sa.String(), nullable=False),
+        sa.Column("encrypted_access_token", sa.String(), nullable=False),
+        sa.Column("encrypted_refresh_token", sa.String(), nullable=True),
+        sa.Column("expires_at", sa.DateTime(timezone=True), nullable=True),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=True,
+        ),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"]),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("user_id", "provider", name="uq_oauth_account_user_provider"),
+    )
+    op.create_index(
+        op.f("ix_oauth_accounts_user_id"), "oauth_accounts", ["user_id"], unique=False
+    )
+
+    # 4. badgesテーブル（issue #31）
+    op.create_table(
+        "badges",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("user_id", sa.Integer(), nullable=False),
+        sa.Column("category", sa.String(), nullable=False),
+        sa.Column("tier", sa.Integer(), nullable=False),
+        sa.Column(
+            "earned_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=True,
+        ),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"]),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("user_id", "category", "tier", name="uq_badge_user_category_tier"),
+    )
+    op.create_index(op.f("ix_badges_user_id"), "badges", ["user_id"], unique=False)
+
+    # 5. questsテーブル（issue #31）
+    op.create_table(
+        "quests",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("title", sa.String(), nullable=False),
+        sa.Column("description", sa.Text(), nullable=False),
+        sa.Column("difficulty", sa.Integer(), nullable=False),
+        sa.Column("category", sa.String(), nullable=False),
+        sa.Column("is_generated", sa.Boolean(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=True,
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+
+    # 6. skill_treesテーブル（issue #31）
+    op.create_table(
+        "skill_trees",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("user_id", sa.Integer(), nullable=False),
+        sa.Column("category", sa.String(), nullable=False),
+        sa.Column("tree_data", sa.JSON(), nullable=False, server_default=sa.text("'{}'")),
+        sa.Column("generated_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"]),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("user_id", "category", name="uq_skill_tree_user_category"),
+    )
+    op.create_index(op.f("ix_skill_trees_user_id"), "skill_trees", ["user_id"], unique=False)
+
+    # 7. quest_progressテーブル（issue #31、usersとquestsに依存）
+    op.create_table(
+        "quest_progress",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("user_id", sa.Integer(), nullable=False),
+        sa.Column("quest_id", sa.Integer(), nullable=False),
+        sa.Column("status", sa.String(), nullable=False),
+        sa.Column(
+            "started_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=True,
+        ),
+        sa.Column("completed_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["quest_id"], ["quests.id"]),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"]),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("user_id", "quest_id", name="uq_quest_progress_user_quest"),
+    )
+    op.create_index(
+        op.f("ix_quest_progress_user_id"), "quest_progress", ["user_id"], unique=False
+    )
+
+
+def downgrade() -> None:
+    """全テーブルを削除（逆順）"""
+    op.drop_index(op.f("ix_quest_progress_user_id"), table_name="quest_progress")
+    op.drop_table("quest_progress")
+
+    op.drop_index(op.f("ix_skill_trees_user_id"), table_name="skill_trees")
+    op.drop_table("skill_trees")
+
+    op.drop_table("quests")
+
+    op.drop_index(op.f("ix_badges_user_id"), table_name="badges")
+    op.drop_table("badges")
+
+    op.drop_index(op.f("ix_oauth_accounts_user_id"), table_name="oauth_accounts")
+    op.drop_table("oauth_accounts")
+
+    op.drop_index(op.f("ix_profiles_user_id"), table_name="profiles")
+    op.drop_table("profiles")
+
+    op.drop_index(op.f("ix_users_username"), table_name="users")
+    op.drop_index(op.f("ix_users_id"), table_name="users")
+    op.drop_table("users")
