--- backend/alembic/versions/0003_create_oauth_accounts_table.py
@@ -1,39 +1,0 @@
-"""create oauth_accounts table
-
-Revision ID: 0003_oauth_accounts
-Revises: 0002_profiles
-Create Date: 2026-02-18
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-from alembic import op
-
-revision: str = "0003_oauth_accounts"
-down_revision: Union[str, Sequence[str], None] = "0002_profiles"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    op.create_table(
-        "oauth_accounts",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("user_id", sa.Integer(), nullable=False),
-        sa.Column("provider", sa.String(), nullable=False),
-        sa.Column("provider_user_id", sa.String(), nullable=False),
-        sa.Column("encrypted_access_token", sa.String(), nullable=False),
-        sa.Column("encrypted_refresh_token", sa.String(), nullable=True),
-        sa.Column("expires_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"]),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("user_id", "provider", name="uq_oauth_user_provider"),
-    )
-    op.create_index(op.f("ix_oauth_accounts_user_id"), "oauth_accounts", ["user_id"], unique=False)
-
-
-def downgrade() -> None:
-    op.drop_index(op.f("ix_oauth_accounts_user_id"), table_name="oauth_accounts")
-    op.drop_table("oauth_accounts")
