--- backend/alembic/versions/0005_create_quests_table.py
@@ -1,35 +1,0 @@
-"""create quests table
-
-Revision ID: 0005_quests
-Revises: 0004_badges
-Create Date: 2026-02-18
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-from alembic import op
-
-revision: str = "0005_quests"
-down_revision: Union[str, Sequence[str], None] = "0004_badges"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    op.create_table(
-        "quests",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("title", sa.String(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=False),
-        sa.Column("difficulty", sa.Integer(), nullable=False),
-        sa.Column("category", sa.String(), nullable=False),
-        sa.Column("is_generated", sa.Boolean(), nullable=False),
-        sa.Column("created_at", sa.DateTime(timezone=True), server_default=sa.text("(CURRENT_TIMESTAMP)"), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-
-
-def downgrade() -> None:
-    op.drop_table("quests")
